requirement.txt
dnspython
colorama
beautifulsoup4
pysocks
requests